refactor: move EVM primitives out of framework into owning domains
