+++ src/features/polymarket/withdrawalConfig.ts
@@ -6,7 +6,7 @@ import { getAddress, type Address } from 'viem';
 import { analytics } from '@/analytics';
 import { USD_DECIMALS } from '@/features/currency/constants';
 import { ChainId } from '@/features/network/types/backendNetworks';
-import { encodeErc20Transfer } from '@/framework/core/evm/erc20Calldata';
+import { encodeErc20Transfer } from '@/features/token/core/services/erc20Calldata';
 import { time } from '@/framework/core/utils/time';
 import { logger, RainbowError } from '@/logger';
 import { USDC_ADDRESS } from '@/references/constants';
